make variable names consistent with data bag
we were using 'apikey' and 'username' in the data bag, but everywhere else calling it 'rackspace_username'
and 'rackspace_api_key', which is inconsistent and a bit confusing. This changes the data bag to use
'rackspace_api_key' and 'rackspace_username' as keys instead of 'username' and 'apikey'. This change as it stands
will currently break existing users unless they change their data bag accordingly