Add depends lines for Chef 11
We use either apt_repository or yum_repository depending on the distro.
If for some reason a user didn't have either the apt or the yum repo in
their runlist prior to this cookbook these [would fail on Chef 11][1].

The apt depends is specifically set to the version after which
apt_repository supported cookbook_file as the key attribute.

[1]: http://docs.opscode.com/breaking_changes_chef_11.html#non-recipe-file-evaluation-includes-dependencies