Raise limit in api calls
There is currently a [bug in Rackspace Monitoring fog][1] where it fails
to handle pagination. This means if your entity is outside of the
default limit of 100 most of this cookbook will fail. It will create
duplicate entities then create duplicate checks on those entities. The
next time it runs it will do the same since the duplicate entity is also
outside the limit.

[1]: https://github.com/fog/fog/issues/2469